pyproject.toml
[tool.poetry]
name = "lights"
version = "0.1.0"
description = "Control an Adafruit NeoPixel light strip from a Raspberry Pi"
authors = ["Alex Krantz <alex@krantz.dev>"]
license = "MIT"
packages = [
    { include = "lights_common" },
    { include = "lights_controller" },
]

[tool.poetry.dependencies]
python = "^3.7"
pycapnp = "^1.1.0"
click = "^8.0.1"
"RPi.GPIO" = "^0.7.0"
Adafruit-Blinka = "^6.13.1"
rpi-ws281x = "^4.3.0"
adafruit-circuitpython-neopixel = "^6.2.1"
python-dotenv = "^0.19.0"
uvicorn = {version = "^0.15.0", extras = ["standard"]}
fastapi = "^0.68.1"
ujson = "^4.2.0"
gunicorn = "^20.1.0"
alembic = "^1.7.3"
asyncpg = "^0.24.0"
SQLAlchemy = {version = "^1.4.25", extras = ["asyncio"]}

[tool.poetry.dev-dependencies]
black = "^21.9b0"
sqlalchemy2-stubs = "^0.0.2-alpha.17"

[tool.poetry.scripts]
lights-controller = "lights_controller.main:main"
lights-web = "lights_web.main:main"

[tool.black]
line-length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
